+++ tests/test_commandline.py
@@ -201,6 +201,12 @@ def test_burned_in_annotation_fail(caplog):
     assert expected_message in str(log_record.exc_info[1])
 
 
+def test_in_place_and_output_directory_are_exclusive():
+    with pytest.raises(SystemExit) as e:
+        run_dicognito(path_to(""), "--output-dir", "some_output_dir", "--in-place")
+        assert "argument --in-place: not allowed with argument --output-directory/-o" in str(e.value)
+
+
 def test_creates_output_directory_when_missing():
     run_dicognito(path_to(""), "--output-dir", path_to("new_dir"))