+++ src/dicognito/release_notes.md
@@ -1,6 +1,7 @@
 ### New
 
-- `--in-place` flag added to explicitly specify in place editing of source files
+- `--in-place` flag added to explicitly specify in place editing of source files. It's an error
+  to specify both `--in-place` and `--output-directory`/`-o`.
   ([#134](https://github.com/blairconrad/dicognito/issues/134))
 
 ## 0.14.0